More truncation fixes.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5251 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Cocoa/MyOpenGLView.m
@@ -650,7 +650,7 @@ FAIL:
 /* Capture shift, ctrl, opt and command press & release */
 - (void)flagsChanged:(NSEvent *)theEvent
 {
-	int flags = [theEvent modifierFlags];
+	NSUInteger flags = [theEvent modifierFlags];
 	opt = (flags & NSAlternateKeyMask) ? YES : NO;
 	ctrl = (flags & NSControlKeyMask) ? YES : NO;
 	command = (flags & NSCommandKeyMask) ? YES : NO;
